+++ polars/polars-lazy/src/physical_plan/expressions/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod column;
 pub mod default;
 mod final_agg;
 pub(crate) mod literal;
+pub(crate) mod sort;
 pub(crate) mod take;
 pub(crate) mod window;
