+++ polars/polars-lazy/src/prelude.rs
@@ -17,6 +17,7 @@ pub use crate::{
 pub(crate) use crate::{
     logical_plan::{aexpr::*, alp::*, conversion::*},
     physical_plan::expressions::{
-        column::ColumnExpr, literal::LiteralExpr, take::TakeExpr, window::WindowExpr,
+        column::ColumnExpr, literal::LiteralExpr, sort::SortExpr, take::TakeExpr,
+        window::WindowExpr,
     },
 };
